test: 원시값 포장 Prize 클래스
